Allow theme assets to be uploaded to CDN
In order to improve page load performance it is a good idea to have static
assets like images, scripts & styles in a CDN rather than being served from
the website. This not only reduces the amount of requests that the website
server has to handle but also means the assets can be distributed across the
world via a CDN.

Additional arguments need to be included in the `grunt dist` task in order to
trigger the uploading of assets to a CDN and update theme files to point to
said CDN. Below is an example of all the required parameters in order to trigger
the upload task.

```
grunt dist -cdnUrl=https://az6491939.vo.msecnd.net -container=theme
-accountName=orcharddevelopment
-accountKey=ysnUsnOAstGaJO9myM2PiCQTUrO9KLE5lQe1P2HIDvDgY6xaLanBMFUHslElTWvNlqRQhuDIJ72iSvEFWe8JNg==
```
It should be noted that the `cdnUrl` argument should not contain a trailing
slash.